+++ src/main/java/com/example/plugissue/station/service/StationService.java
@@ -21,7 +21,7 @@ public class StationService {
 
     private final StationRepository stationRepository;
 
-    public List<StationStatusDto> findStationsStatusByLoc(Double lat, Double lng, Integer range) {
+    public List<StationStatusDto> findStationsStatusByLoc(Double lat, Double lng, Double range) {
         List<Object[]> queryResult = stationRepository.findStationsByLoc(lat, lng, range);
         List<StationStatusDto> dtos = new ArrayList<>();
